moving template compiler construction into app
